Scenario with Charactet that has armor

+++ Assets/Scripts/Inventory.cs
@@ -24,7 +24,7 @@ public class Inventory
         return null;
     }
 
-    public float GetTotalArmor()
+    public int GetTotalArmor()
     {
         int totalArmor = equippedItems.Values.Sum(t => t.Armor);
         return totalArmor;
